+++ src/cpu/x64/jit_uni_binary_kernel.cpp
@@ -143,8 +143,8 @@ void jit_uni_binary_kernel_t<isa, Vmm>::init_post_ops_injector() {
     const binary_injector::rhs_arg_static_params_t rhs_arg_bsp {10, reg_tmp_,
             reg_elt_inj_table_, r13, true /*preserve gpr*/,
             true /*preserve vmm*/, PARAM_OFF(post_ops_binary_rhs_arg_vec),
-            PARAM_OFF(dst_orig), dst_d, tail_size_, tail_opmask_,
-            false /*use_exact_tail_scalar_bcast*/};
+            PARAM_OFF(dst_orig), dst_d, static_cast<int>(tail_size_),
+            tail_opmask_, false /*use_exact_tail_scalar_bcast*/};
     const binary_injector::static_params_t bsp(this->param1,
             get_supported_postops_bcast_strategies(), rhs_arg_bsp);
