+++ src/cpu/x64/jit_uni_pool_kernel.cpp
@@ -89,14 +89,13 @@ jit_uni_pool_kernel_t<isa>::jit_uni_pool_kernel_t(
         static constexpr bool use_exact_tail_scalar_bcast = false;
 
         const binary_injector::rhs_arg_static_params_t rhs_sp {
-                static_cast<std::size_t>(this->xmm4.getIdx()), this->r14,
-                this->r15, this->r13, preserve_gpr, preserve_vmm,
+                this->xmm4.getIdx(), this->r14, this->r15, this->r13,
+                preserve_gpr, preserve_vmm,
                 GET_OFF(post_ops_binary_rhs_arg_vec), GET_OFF(dst_orig),
                 memory_desc_wrapper(jpp.tag_kind == jit_memory_tag_kind_t::ncsp
                                 ? jpp.tmp_md
                                 : *dst_md),
-                static_cast<size_t>(tail_size), k_c_tail_mask,
-                use_exact_tail_scalar_bcast};
+                tail_size, k_c_tail_mask, use_exact_tail_scalar_bcast};
 
         const binary_injector::static_params_t bsp {reg_param,
                 get_supported_bcast_strategies(), rhs_sp, f8_e5m2_cvt_.get(),
